Pequeña correción de typecast

+++ trunk/BasicEngine/BasicEngine/Graphics/Fonts/acgfx_font.cpp
@@ -903,7 +903,7 @@ void CFontLoaderTextFormat::InterpretChar(string &str, int start)
 			chnl = strtol(value.c_str(), 0, 10);
 
 		//if( pos == str.size() ) break;  //orig Jesús
-		if( pos2 >= str.size() -1 ) break;  //Yorman  y José Manuel (añadido)
+		if( pos2 >= (int)str.size() -1 ) break;  //Yorman  y José Manuel (añadido)
 	}
 
 	// Store the attributes
@@ -984,7 +984,7 @@ void CFontLoaderTextFormat::InterpretInfo(string &str, int start)
 			outlineThickness = (short)strtol(value.c_str(), 0, 10);
 
 		//if( pos >= str.size() ) break; //orig Jesús
-		if( pos2 >= str.size() -1) 	break; //Yorman  y José Manuel (añadido)
+		if( pos2 >= (int)str.size() -1) 	break; //Yorman  y José Manuel (añadido)
 
 
 	}
@@ -1020,7 +1020,7 @@ void CFontLoaderTextFormat::InterpretPage(string &str, int start, const char *fo
 			file = value.substr(1, value.length()-2);
 
 		//if( pos == str.size() ) break; //orig Jesús
-		if( pos2 >= str.size() - 1) break; //Yorman  y José Manuel (añadido)
+		if( pos2 >= (int)str.size() - 1) break; //Yorman  y José Manuel (añadido)
 	}
 
 	LoadPage(id, file.c_str(), fontFile);
